AU-1220: refactor code to base
